refactor: unify types and distance_metric modules into a single unified traits modules to make things a bit neater, especially in the docs
