+++ src/float/query/nearest_one.rs
@@ -1,11 +1,11 @@
 use az::{Az, Cast};
 use std::ops::Rem;
 
-use crate::distance_metric::DistanceMetric;
 use crate::float::kdtree::{Axis, KdTree, LeafNode};
 use crate::generate_nearest_one;
 use crate::nearest_neighbour::NearestNeighbour;
-use crate::types::{is_stem_index, Content, Index};
+use crate::traits::DistanceMetric;
+use crate::traits::{is_stem_index, Content, Index};
 
 macro_rules! generate_float_nearest_one {
     ($leafnode:ident, $doctest_build_tree:tt) => {
@@ -76,10 +76,10 @@ where
 
 #[cfg(test)]
 mod tests {
-    use crate::distance_metric::DistanceMetric;
     use crate::float::distance::Manhattan;
     use crate::float::kdtree::{Axis, KdTree};
     use crate::nearest_neighbour::NearestNeighbour;
+    use crate::traits::DistanceMetric;
     use rand::Rng;
 
     type AX = f32;
