+++ src/float_leaf_slice/fallback.rs
@@ -2,7 +2,7 @@ use az::Cast;
 use std::collections::BinaryHeap;
 
 use crate::float::result_collection::ResultCollection;
-use crate::{float::kdtree::Axis, types::Content, BestNeighbour, NearestNeighbour};
+use crate::{float::kdtree::Axis, traits::Content, BestNeighbour, NearestNeighbour};
 
 #[inline]
 pub(crate) fn update_nearest_dist_autovec<A: Axis, T: Content>(
