+++ src/immutable/float/query/nearest_n_within.rs
@@ -4,13 +4,13 @@ use std::collections::BinaryHeap;
 use std::num::NonZero;
 use std::ops::Rem;
 
-use crate::distance_metric::DistanceMetric;
 use crate::float::kdtree::Axis;
 use crate::float::result_collection::ResultCollection;
 use crate::float_leaf_slice::leaf_slice::{LeafSliceFloat, LeafSliceFloatChunk};
 use crate::immutable::float::kdtree::ImmutableKdTree;
 use crate::nearest_neighbour::NearestNeighbour;
-use crate::types::Content;
+use crate::traits::Content;
+use crate::traits::DistanceMetric;
 
 use crate::generate_immutable_nearest_n_within;
 
@@ -83,10 +83,10 @@ where
 
 #[cfg(test)]
 mod tests {
-    use crate::distance_metric::DistanceMetric;
     use crate::float::distance::SquaredEuclidean;
     use crate::float::kdtree::Axis;
     use crate::immutable::float::kdtree::ImmutableKdTree;
+    use crate::traits::DistanceMetric;
     use rand::Rng;
     use std::cmp::Ordering;
     use std::num::NonZero;
