+++ src/immutable/float/query/nearest_one.rs
@@ -1,13 +1,13 @@
 use az::Cast;
 use std::ops::Rem;
 
-use crate::distance_metric::DistanceMetric;
 use crate::float::kdtree::Axis;
 use crate::float_leaf_slice::leaf_slice::{LeafSliceFloat, LeafSliceFloatChunk};
 use crate::generate_immutable_nearest_one;
 use crate::immutable::float::kdtree::ImmutableKdTree;
 use crate::nearest_neighbour::NearestNeighbour;
-use crate::types::Content;
+use crate::traits::Content;
+use crate::traits::DistanceMetric;
 
 macro_rules! generate_immutable_float_nearest_one {
     ($doctest_build_tree:tt) => {
@@ -74,11 +74,11 @@ where
 
 #[cfg(test)]
 mod tests {
-    use crate::distance_metric::DistanceMetric;
     use crate::float::distance::SquaredEuclidean;
     use crate::float::kdtree::Axis;
     use crate::immutable::float::kdtree::ImmutableKdTree;
     use crate::nearest_neighbour::NearestNeighbour;
+    use crate::traits::DistanceMetric;
     use rand::{Rng, SeedableRng};
 
     #[test]
