+++ src/immutable/float/query/within.rs
@@ -1,10 +1,10 @@
-use crate::distance_metric::DistanceMetric;
 use crate::float::kdtree::Axis;
 use crate::float_leaf_slice::leaf_slice::{LeafSliceFloat, LeafSliceFloatChunk};
 use crate::generate_immutable_within;
 use crate::immutable::float::kdtree::ImmutableKdTree;
 use crate::nearest_neighbour::NearestNeighbour;
-use crate::types::Content;
+use crate::traits::Content;
+use crate::traits::DistanceMetric;
 use az::Cast;
 
 macro_rules! generate_immutable_float_within {
@@ -71,10 +71,10 @@ impl<
 
 #[cfg(test)]
 mod tests {
-    use crate::distance_metric::DistanceMetric;
     use crate::float::distance::Manhattan;
     use crate::float::kdtree::Axis;
     use crate::immutable::float::kdtree::ImmutableKdTree;
+    use crate::traits::DistanceMetric;
     use rand::Rng;
     use std::cmp::Ordering;
