+++ src/test_utils.rs
@@ -12,7 +12,7 @@ use crate::fixed::kdtree::{Axis as AxisFixed, KdTree as FixedKdTree};
 use crate::float::kdtree::{Axis, KdTree};
 use crate::float_leaf_slice::leaf_slice::{LeafSliceFloat, LeafSliceFloatChunk};
 use crate::immutable::float::kdtree::ImmutableKdTree;
-use crate::types::{Content, Index};
+use crate::traits::{Content, Index};
 
 // use rand_distr::UnitSphere as SPHERE;
